requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
Jinja2>=3.1.0
